Icons components have their own folder now
